Refactor core aicenter processing logic
